Fix some cases where heaps larger than 2G would wrap because of signed 32-bit variables. --- Add debug mode to force a large heap.
I ran full regressions with:
_BK_HEAP_ADDJUNK=0x90000000 make -j6 -k

The handful of failures were all found to be expected.
And t.compat-6.0 runs forever because bk-6 is missing
this RTI.

bk: 55080624TokPqwnFl5-rY28wqCb0VA

+++ src/sccs.h
@@ -1567,7 +1567,7 @@ u32	sccs_hasRootkey(sccs *s, char *key);
 void	sccs_loadHeapMeta(sccs *s);
 typedef	struct MAP MAP;
 void	*dataAlloc(u32 esize, u32 nmemb);
-void	datamap(char *name, void *start, int len,
+void	datamap(char *name, void *start, size_t len,
     FILE *f, long off, int byteswap, int *didpage);
 void	dataunmap(FILE *f, int keep);
 FILE	*fopen_bkfile(char *file, char *mode, u64 size, int chkxor);
